fix using different location other than storage
Moving pdfs to a different location other than the default zotero
storage results in unable to find pdf. Therefore change the way in which
the file path to pdfs is created when using a file path other than
storage is used. I create some debug statements in case it doesn't work
for the default location. I am unable to test if this fix works for the
default location or not.