'Add absolute path to bib location'
Simple helper function to join the parsed bib from YAML to the path of
the current buffer. This is based on the assumption that for pandoc to
work it must reference the bib location must be given relative to the
current file.

Hopefully, fixes the issue where files are opened by nvim from a remote
location (say from ~ with a fuzzy finder).

TODO: Only implemented for quarto, consider latex as well.