fixed the highlight order and added option to turn headings on or off
Highlight grouping was previously done based off the alphabetical order
of the headings, this is wrong. We want the order of the highlights to
be by the color of the highlight themselves which doesn't change in
zotero (only the yellow highlight might be different in zotero 6 vs 7).

---

Also added option to turn off headings altogether since the user may not
want headings.